Update T-Stick_2GW_Connecting_Guide(v1.2).md
I found that my config page differed from that shown, and I needed to instead reboot my T-Stick. I was subsequently unable to follow the rest of the guide, since I didn't know the stick's new IP address, but since it connected to libmapper at this point anyways it wasn't an issue.